Correctly handle the outputting of heartbeat Sets in GenericMaker (#116)
We were previously sending out heartbeat Set<T>s as soon as we received
them, but that was wrong, because the output buffer was potentially
holding Ts with times earlier than the heartbeat's time. Those Ts were
then output later than the heartbeat, breaking the ordering guarantee.

So now, TimeSliceOutputBuffer holds _two_ buffers: the buffer of
Ts (which it had before) and a new buffer of heartbeat
Set<T>s. Heartbeat Set<T>s are added via the buffer_heartbeat()
function. Adding a heartbeat to the buffer advances the latest seen
time, which is used to determine what data is ready to be sent
out. When flush() is called, the return (by reference) value is the
next heartbeat Set<T> if it sits at the start of the current time
window; otherwise it is a payload Set<T> with the Ts that fall in the
current window.