In TXBuffer, make receive timeouts 0 so we can handle high rates
We were previously waiting on the receives from the TX input buffer
and on the data request input buffer. But then if the input rate from
one of queues (typically TPs) was greater than 1/(receive timeout),
the unconditional wait on the other queue would cause us to back
up. We now only wait if neither input queue received anything.