Buffer TPs, TAs and TCs using the readoutlibs buffer implementation (#110)
* Initial TPBuffer, using readoutlibs' buffer/data request interface

Doesn't compile yet

* Add boost::iostreams dependency

This is really a dependency of readoutlibs, not trigger, but I had to
put something here to make it build

* Actually implement TPWrapper::set_first_timestamp()

This turns out to be needed by the search functionality in DefaultRequestHandlerModel

* Implement do_work function

* Add schema file for configuration info

* Add TABuffer

Currently just a copy-paste of TPBuffer with appropriate
changes. Could consider doing this with templating

* Remove unused #includes

* Implement get_frame_iterator_timestamp for TPs

* Add schema for TABuffer

* Replace tpbuffer.jsonnet and tabuffer.jsonnet with txbuffer.jsonnet

An epiphany: the configs for TPBuffer and TABuffer are the same, so we
only need one of them

* Bugfix: serialize the TA and return serialized form in begin()/end()

This ensures that the TP list in the TA is also saved to the TR

* Add TCBuffer

* Remove unused get_timestamp() function

* Add TriggerCandidateNQ so we can send TCs over the network

This is currently needed for the hack that allows us to fan out TCs to
MLT and the TC buffer, in which we send TCs over a pub/sub network
connection from the trigger app back into the trigger app

* Make FakeTPCreatorHeartbeatMaker less verbose

* Add some debugging output

* Add find_dependency(readoutlibs)

Thanks for Eric for pointing out the missing line

* Adapt to changes in readoutlibs for partial fragment on timeout

* Add print_ds_fragments test app

print_ds_fragments will print a text representation of the DS
fragments in an HDF5 file. The purpose is as a spot check that the
TP/TA/TC buffering and readout is working correctly.

* Remove now-unused fragment output queue

* Replace QueueTimeoutExpired with TimeoutExpired

* Add Tee module, which duplicates its input on two output queues

This is useful for duplicating TP/TA/TC to the trigger chain and
readout buffers

* Make TXBuffer use new communications API

* Fix queue name

* Update an output warning

* Add get_name() to a TLOG() for benefit of debugging

* Fix unused-variable compiler warnings

* When send fails, print _which_ queue the send was to

* In TXBuffer, make receive timeouts 0 so we can handle high rates

We were previously waiting on the receives from the TX input buffer
and on the data request input buffer. But then if the input rate from
one of queues (typically TPs) was greater than 1/(receive timeout),
the unconditional wait on the other queue would cause us to back
up. We now only wait if neither input queue received anything.