Output heartbeats based on time passed, not just on payload receipt
We were previously only outputting heartbeats when a payload TPSet was
received. This commit changes the logic so that we estimate the
current timestamp based on the latest payload seen and the wallclock
time. This means that when the input stream of TPSets stops, we
continue sending heartbeats. Also, in the unlikely event that the rate
of input TPSets is very low, we would continue sending heartbeats in
between payload TPSets.